requirements.txt
beautifulsoup4==4.11.2
boto3==1.26.89
botocore==1.29.89
certifi==2022.12.7
charset-normalizer==3.1.0
filelock==3.9.0
idna==3.4
jmespath==1.0.1
lxml==4.9.2
neo4j==5.6.0
praw==7.7.0
prawcore==2.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
snscrape==0.6.1.20230314
soupsieve==2.4
urllib3==1.26.15
